switched out c-style types for zig-native types

+++ src/utils.zig
@@ -1,22 +1,22 @@
-pub fn delay(time: c_ulong) void {
+pub fn delay(time: usize) void {
     var i: usize = time;
     while (i != 0) {
         i = i - 1;
     }
 }
 
-pub fn read32(addr: c_ulong) c_uint {
-    const ptr = @as(*volatile u32, @ptrFromInt(addr));
+pub fn read32(addr: usize) i32 {
+    const ptr = @as(*volatile i32, @ptrFromInt(addr));
     return ptr.*;
 }
 
-pub fn write32(addr: c_ulong, value: c_uint) void {
-    const ptr = @as(*volatile u32, @ptrFromInt(addr));
+pub fn write32(addr: usize, value: i32) void {
+    const ptr = @as(*volatile i32, @ptrFromInt(addr));
     ptr.* = value;
 }
 
-pub fn getEl() c_int {
-    var el: c_int = 0;
+pub fn getEl() i32 {
+    var el: i32 = 0;
 
     asm volatile (//syscall
         "mrs %[result], CurrentEL"
